fix: switch to dynamic imports, more debug
